requirements.txt
astroid==3.3.5
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dill==0.3.9
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pylint==3.3.1
python-dotenv==1.0.1
requests==2.32.3
tomlkit==0.13.2
urllib3==2.2.3
